requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
python-dotenv
langchain
langchain-community
requests
anthropic
fastapi-mcp